dev/pyproject.toml
[project]
name = "agent-workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.1.0",
    "fastapi[standard]>=0.115.4",
    "httpx>=0.27.2",
    "langchain-community>=0.3.5",
    "langchain-openai>=0.2.6",
    "langgraph-checkpoint-sqlite>=2.0.1",
    "langgraph>=0.2.45",
    "lxml>=5.3.0",
    "numexpr>=2.10.1",
    "python-dotenv>=1.0.1",
    "six>=1.16.0",
    "uvicorn>=0.32.0",
    "grandalf>=0.8",
    "beautifulsoup4>=4.12.3",
    "structlog>=24.4.0",
    "jinja2>=3.1.4",
    "rich>=13.9.4",
    "pandas>=2.2.3",
    "matplotlib>=3.10.1",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
